BugFix/Ability_Add/Balance

+++ Assets/1.Main/2.Script/Controller/Player/PlayerSkillController.cs
@@ -23,7 +23,7 @@ public class PlayerSkillController : MonoBehaviour
     float ArmorPierce;
     float Remove;
     int Drain;
-    int Crush;
+    float Crush;
     int Burn;
     #endregion
